Mixture PDFs as chapter 10.3.
I use a non-object-oriented way to write it, which is not aligned with the book. Designing a helper function for this is also difficult. In the end, I found that writing that code in ray_color function might be a simpler way of doing this.

I modify the code that checks if pdf_value is 0 to a more straight-forward statement, and a comment for that is also added.